+++ cryptography/src/test/scala/akka/stream/alpakka/cryptography/scaladsl/CryptographicFlowSpec.scala
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2017 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.cryptography.impl
+package akka.stream.alpakka.cryptography.scaladsl
 
 import java.security.{KeyPair, KeyPairGenerator}
 import javax.crypto.{KeyGenerator, SecretKey}
@@ -10,7 +10,7 @@ import javax.crypto.{KeyGenerator, SecretKey}
 import akka.NotUsed
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
-import akka.stream.alpakka.cryptography.impl.CryptographicFlows._
+import akka.stream.alpakka.cryptography.scaladsl.CryptographicFlows._
 import akka.stream.scaladsl.{Sink, Source}
 import akka.util.ByteString
 import org.scalacheck.{Arbitrary, Gen}
